+++ logger/go.mod
@@ -3,7 +3,10 @@ module github.com/NeverDefineUs/vitaes/logger
 go 1.12
 
 require (
+	github.com/go-redis/redis v6.15.2+incompatible
 	github.com/gorilla/mux v1.7.2
 	github.com/mattn/go-sqlite3 v1.10.0
+	github.com/onsi/ginkgo v1.8.0 // indirect
+	github.com/onsi/gomega v1.5.0 // indirect
 	github.com/rs/cors v1.6.0
 )
